Delete VCR cassettes dir when re-recording
The option to record all doesn't do what I'd expected. In this case when we want to re-record its best to just delete all cassettes to ensure we record again cleanly.

Also for the token, don't reuse an existing cassette as that wasn't playing well with this change. Since its lazily loaded, when the token and client lets are called, the request and response get recorded in the current cassette which we want. We though we could reuse an existing cassette here so we don't have to keep signing in every time but not able to get it to work reliably.