Remove activesupport dev dep for tests
We used this because its nice to have some of the predicate methods like `present` or `in` available for use in tests but we'd rather not have the dependency due to its Ruby version lock in. Plus we have good ways now of confirming test data by using regex matches which is preferrable.