Re-record all tests to use new auth API URL
The new URL is now auth.api.zesty.io so we had to re-record all tests to use it. I re-organized the auth cassettes so its in an auth folder to match API. Re-recording introduced one failure with get instance test since it seems now that screenshot_url has a value by default. This may change though so we may want to rewrite spec so its blank or present.